fix code rewiew issues
