+++ osa_tool/operations/analysis/vkr_scoring/vkr_scorer.py
@@ -15,10 +15,7 @@ Steps:
 from __future__ import annotations
 
 import json
-import os
-import re
 import sys
-from datetime import datetime, timezone
 from pathlib import Path
 from typing import Optional
 
@@ -27,192 +24,9 @@ from osa_tool.core.git.git_agent import GitAgent
 from osa_tool.core.llm.llm import ModelHandlerFactory
 from osa_tool.utils.logger import logger
 
-from .checks import VkrConfig, build_file_tree, run_all_checks
-from .claims import extract_claims, verify_claims
-
-# ── Scoring weights ───────────────────────────────────────────────────────────
-
-WEIGHTS_DATA_EXPERIMENT: dict[str, int] = {
-    "readme": 20,
-    "license": 10,
-    "commits": 10,
-    "execution_files": 20,
-    "requirements": 10,
-    "data_files": 10,
-    "experiment_scripts": 20,
-}
-
-WEIGHTS_APPS: dict[str, int] = {
-    "readme": 25,
-    "license": 10,
-    "commits": 10,
-    "execution_files": 25,
-    "tests": 30,
-}
-
-_APP_TYPES = {"app"}
-
-
-def compute_score(checks: dict, repo_type: str = "unknown") -> tuple[int, dict]:
-    """Return (score_0_100, per_check_breakdown)."""
-    weights = WEIGHTS_APPS if repo_type in _APP_TYPES else WEIGHTS_DATA_EXPERIMENT
-
-    earned = 0
-    max_points = 0
-    breakdown: dict = {}
-
-    for name, weight in weights.items():
-        result = checks.get(name, {})
-        if result.get("applicable") is False:
-            breakdown[name] = {"applicable": False}
-            continue
-
-        if name == "readme":
-            passed = bool(result.get("present") and result.get("meaningful"))
-        else:
-            passed = bool(result.get("present"))
-
-        pts = weight if passed else 0
-        earned += pts
-        max_points += weight
-        breakdown[name] = {"weight": weight, "earned": pts, "passed": passed}
-
-    return earned, breakdown
-
-
-# ── Report building ───────────────────────────────────────────────────────────
-
-_REPO_TYPE_LABELS = {
-    "app": "Application",
-    "algorithm_experiments": "Algorithm Experiments",
-    "model_training_experiments": "Model Training Experiments",
-}
-
-_CHECK_ORDER = [
-    "readme",
-    "license",
-    "commits",
-    "requirements",
-    "execution_files",
-    "repo_type",
-    "tests",
-    "data_files",
-    "experiment_scripts",
-    "syntax",
-    "docstrings",
-]
-
-
-def build_report(checks: dict, repo_url: str) -> dict:
-    repo_type = checks.get("repo_type", {}).get("value", "unknown")
-    score, breakdown = compute_score(checks, repo_type)
-    return {
-        "repo_url": repo_url,
-        "analyzed_at": datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "checks": checks,
-        "summary": {
-            "score": score,
-            "score_breakdown": breakdown,
-            "repo_type": checks.get("repo_type", {}).get("value", "unknown"),
-            "syntax": checks.get("syntax", {}),
-            "docstrings": checks.get("docstrings", {}),
-        },
-    }
-
-
-def _format_check_line(key: str, val: dict) -> str:
-    if key == "repo_type":
-        label = _REPO_TYPE_LABELS.get(val.get("value", ""), val.get("value", "unknown"))
-        conf = val.get("confidence", "")
-        return f"  repo_type       : {label} ({conf} confidence)"
-
-    if val.get("applicable") is False:
-        return f"  {key:<18}: n/a"
-
-    if key == "readme":
-        if not val.get("present"):
-            status = "MISSING"
-        elif val.get("meaningful") is False:
-            status = "present (too short)"
-        else:
-            chars = val.get("char_count", 0)
-            status = f"OK ({chars} chars)"
-        return f"  readme          : {status}"
-
-    if key == "commits":
-        count = val.get("count", 0)
-        status = f"OK (>{5})" if val.get("present") else f"FAIL ({count} commits)"
-        return f"  commits         : {status}"
-
-    if key == "syntax":
-        return f"  syntax          : {val.get('summary', 'unknown')}"
-
-    if key == "docstrings":
-        return f"  docstrings      : {val.get('summary', 'unknown')}"
-
-    status = "OK" if val.get("present") else "MISSING"
-    matched = val.get("matched_file") or ", ".join(val.get("verified", [])[:2])
-    detail = f" [{matched}]" if matched else ""
-    return f"  {key:<18}: {status}{detail}"
-
-
-def build_text_report(report: dict) -> str:
-    checks = report["checks"]
-    score = report["summary"]["score"]
-    repo_type = report["summary"]["repo_type"]
-    type_label = _REPO_TYPE_LABELS.get(repo_type, repo_type)
-
-    lines = [
-        f"Repository : {report['repo_url']}",
-        f"Analyzed   : {report['analyzed_at']}",
-        f"Type       : {type_label}",
-        "",
-        "Checks:",
-    ]
-    for key in _CHECK_ORDER:
-        if key in checks:
-            lines.append(_format_check_line(key, checks[key]))
-
-    lines += ["", f"Score: {score}/100"]
-
-    ca = report.get("claims_analysis")
-    if ca:
-        stats = ca.get("stats", {})
-        lines += [
-            "",
-            "Claims Analysis:",
-            f"  Total claims    : {stats.get('total', 0)}",
-            f"  Implemented     : {stats.get('implemented', 0)}",
-            f"  Implementation  : {stats.get('implementation_rate_pct', 0)}%",
-        ]
-
-    return "\n".join(lines)
-
-
-def _sanitize_dir_name(repo_url: str) -> str:
-    name = repo_url.rstrip("/")
-    if "github.com" in name:
-        name = name.split("github.com/", 1)[1]
-    name = name.replace("/", "__")
-    return re.sub(r"[^\w\-.]", "_", name)[:100]
-
-
-def save_results(report: dict, output_dir: str) -> tuple[str, str]:
-    target = os.path.join(output_dir, _sanitize_dir_name(report["repo_url"]))
-    os.makedirs(target, exist_ok=True)
-
-    json_path = os.path.join(target, "report.json")
-    txt_path = os.path.join(target, "report.txt")
-
-    with open(json_path, "w", encoding="utf-8") as f:
-        json.dump(report, f, ensure_ascii=False, indent=2)
-    with open(txt_path, "w", encoding="utf-8") as f:
-        f.write(build_text_report(report))
-
-    return json_path, txt_path
-
-
-# ── Main scorer class ─────────────────────────────────────────────────────────
+from .checks import VkrChecker, VkrConfig, build_file_tree
+from .claims import ClaimsPipeline
+from .scoring_engine import ScoringEngine
 
 
 class VkrScorer:
@@ -225,6 +39,8 @@ class VkrScorer:
         paper_path: Optional[str] = None,
         output_dir: Optional[str] = None,
     ):
+        import os
+
         self._config_manager = config_manager
         self._git_agent = git_agent
         self._paper_path = paper_path
@@ -250,27 +66,30 @@ class VkrScorer:
         config = self._vkr_config
         logger.info(f"VKR quality checks: {config.repo_url}")
         flat_paths, all_paths = build_file_tree(config.clone_dir)
-        checks = run_all_checks(flat_paths, all_paths, config)
-        return build_report(checks, config.repo_url)
+        checks = VkrChecker(config).run_all(flat_paths, all_paths)
+        return ScoringEngine(config.repo_url).build_report(checks)
 
     def run(self) -> dict:
         config = self._vkr_config
+        scorer = ScoringEngine(config.repo_url)
+
         logger.info(f"VKR scoring: {config.repo_url}")
 
         logger.info("Building file tree from local clone...")
         flat_paths, all_paths = build_file_tree(config.clone_dir)
 
         logger.info("Running quality checks...")
-        checks = run_all_checks(flat_paths, all_paths, config)
-        report = build_report(checks, config.repo_url)
+        checks = VkrChecker(config).run_all(flat_paths, all_paths)
+        report = scorer.build_report(checks)
 
         paper_sections = self._load_paper_sections()
         if paper_sections:
+            pipeline = ClaimsPipeline(config)
             logger.info("Extracting claims from paper...")
-            claims = extract_claims(paper_sections, config)
+            claims = pipeline.extract(paper_sections)
             if claims:
                 logger.info(f"Verifying {len(claims)} claims against repository...")
-                report["claims_analysis"] = verify_claims(claims, flat_paths, config)
+                report["claims_analysis"] = pipeline.verify(claims, flat_paths)
             else:
                 logger.warning("No claims extracted from paper.")
                 report["claims_analysis"] = {
@@ -283,11 +102,11 @@ class VkrScorer:
                     },
                 }
 
-        json_path, txt_path = save_results(report, self._output_dir)
+        json_path, txt_path = scorer.save_results(report, self._output_dir)
         logger.info(f"VKR report saved: {json_path}")
         logger.info(f"               : {txt_path}")
 
-        print("\n" + build_text_report(report), file=sys.stderr)
+        print("\n" + scorer.build_text_report(report), file=sys.stderr)
 
         return {
             "result": {
