+++ tests/unit/operations/analysis/vkr_scoring/test_claims.py
@@ -5,29 +5,25 @@ from __future__ import annotations
 import pytest
 
 from osa_tool.utils.response_cleaner import JsonParseError
-from osa_tool.operations.analysis.vkr_scoring.claims import (
-    _parse_json_list,
-    _candidate_files,
-    _truncate,
-)
+from osa_tool.operations.analysis.vkr_scoring.claims import ClaimsPipeline
 
 # ── _parse_json_list ──────────────────────────────────────────────────────────
 
 
 def test_parse_json_list_valid():
-    result = _parse_json_list('[{"a": 1}]')
+    result = ClaimsPipeline._parse_json_list('[{"a": 1}]')
     assert isinstance(result, list)
     assert result[0]["a"] == 1
 
 
 def test_parse_json_list_invalid_json():
     with pytest.raises(JsonParseError):
-        _parse_json_list("not valid json {{{")
+        ClaimsPipeline._parse_json_list("not valid json {{{")
 
 
 def test_parse_json_list_not_a_list():
     with pytest.raises(JsonParseError):
-        _parse_json_list('{"key": "val"}')
+        ClaimsPipeline._parse_json_list('{"key": "val"}')
 
 
 # ── _candidate_files ──────────────────────────────────────────────────────────
@@ -41,7 +37,7 @@ def test_candidate_files_matches_patterns():
         "README.md",
         "setup.py",
     ]
-    result = _candidate_files(filenames)
+    result = ClaimsPipeline._candidate_files(filenames)
     assert "train.py" in result
     assert "model.py" in result
     assert "data.py" in result
@@ -55,7 +51,7 @@ def test_candidate_files_matches_patterns():
 def test_truncate_long_text():
     lines = [f"line {i}" for i in range(300)]
     text = "\n".join(lines)
-    result = _truncate(text, max_lines=250)
+    result = ClaimsPipeline._truncate(text, max_lines=250)
     result_lines = result.splitlines()
     assert len(result_lines) <= 251  # 250 lines + truncation note
     assert "truncated" in result
@@ -64,5 +60,5 @@ def test_truncate_long_text():
 def test_truncate_short_text():
     lines = [f"line {i}" for i in range(10)]
     text = "\n".join(lines)
-    result = _truncate(text, max_lines=250)
+    result = ClaimsPipeline._truncate(text, max_lines=250)
     assert result == text
